requirements.txt
pytest==7.1.2
requests==2.27.1
discord.py==1.7.3
SQLAlchemy==1.4.36
python-dotenv==0.20.0